+++ messageRepositories/ustrirovmessagerepository.h
@@ -23,7 +23,7 @@ public:
     DevicesAdjustingKitMessage &GetMessage();
 public:
     const quint16 GetDistanceToLocator() const;
-    const quint16 GetFvco() const;
+    const quint32 GetFvco() const;
 private:
     void ResetRepository();
